test(taiko-client-rs): dedupe and extract driver event tests
