+++ packages/taiko-client-rs/crates/driver/src/lib.rs
@@ -23,6 +23,9 @@ pub mod preconf_ingress_sync;
 pub mod production;
 /// Synchronization stages and event scanning.
 pub mod sync;
+/// Shared test fixtures for driver unit tests.
+#[cfg(test)]
+pub(crate) mod test_support;
 
 pub use config::DriverConfig;
 pub use driver::Driver;
